Refactor AllowCIDRMiddleware to non-legacy pattern
This moves us away from using django.utils.deprecation.MiddlewareMixin for
long-obsolete versions of Django, and unlocks Django 4.0 compatibility

Also updates Tox testing matrix in line with the shift and adds viable config
to run each tox env via a GH Action. It's not the DRYest, but it does mean
we call each Tox env, rather than the 'best fit' env for the given version
of Python, which only tests the highest compatible version of Django. There
will likely be a cleaner way to do this, which we'll add in the future

Broadening tox-powered coverage