Reinstate automatic publishing to PyPI
This changeset uses a Github Action to build a new distribution and upload it
to PyPI, if the relevant triggering commit is tagged with a tag that matches
the X.Y.Z format.

It uses an allowed-by-MozIT third-party GH Action and the workflow spec drew
plenty of inspiration from its use elsewhere at Mozilla
(github.com/mozilla/mozilla-django-oidc).

This changeset also amends the CI workflow slightly to make it reusable, so
it can be called beforehand by the new Release workflow.

A new, project-specific token for PyPI upload has already been added as a
repository secret.