Reinstate automatic publishing to PyPI (#24)
* Reinstate automatic publishing to PyPI

This changeset uses a Github Action to build a new distribution and upload it
to PyPI, if the relevant triggering commit is tagged with a tag that matches
the X.Y.Z format.

It uses an allowed-by-MozIT third-party GH Action and the workflow spec drew
plenty of inspiration from its use elsewhere at Mozilla
(github.com/mozilla/mozilla-django-oidc).

This changeset also amends the CI workflow slightly to make it reusable, so
it can be called beforehand by the new Release workflow.

A new, project-specific token for PyPI upload has already been added as a
repository secret.

* Simplify package-building syntax

* Refactor GH Action workflow to put the release step into main CI, triggered conditionally

Cutting a new release on GH will trigger a new build to be pushed to PyPI

* Remove redundant workflow_call trigger

* Tidy README

* Update workflow to only run on pushes to master (only), or on pull requests

...rather than pushes to any branch OR pull requests, which caused double
execution of the workflow